Complete task
Add a try-catch block for operator new in BFS function.

+++ Ternary tree/Ternary tree/main.cpp
@@ -18,6 +18,9 @@ struct Node {
     Node* right;
 };
 
+// Function printLevel is used as a helper function in BFS function below.
+// Queue currLevel stores the vertices from the current level and if any of them
+// has a child, it is pushed to queue nextLevel.
 void printLevel(std::queue<Node*>& currLevel, std::queue<Node*>& nextLevel, bool* visited)
 {
     while (!currLevel.empty()) {
@@ -43,10 +46,15 @@ void printLevel(std::queue<Node*>& currLevel, std::queue<Node*>& nextLevel, bool
     std::cout << '\n';
 }
 
+// BFS algorithm is used to read the data of the tree nodes level by level.
 void BFS(Node*& root, unsigned verticesCount) {
-//    if (verticesCount > 0)
-//        return;
-    bool* visited = new bool[verticesCount]; // Exception thrown by new?
+    bool* visited = nullptr;
+    try {
+        visited = new bool[verticesCount];
+    } catch (const std::bad_alloc& ba) {
+        std::cerr << "Bad alloc caught: " << ba.what() << 'n';
+        return;
+    }
     for (size_t i = 0; i < verticesCount; ++i) {
         visited[i] = false;
     }
@@ -61,9 +69,16 @@ void BFS(Node*& root, unsigned verticesCount) {
         printLevel(verticesNextLevel, verticesCurrLevel, visited);
     }
     
-    delete [] visited;
+    delete[] visited;
 }
 
+
+// Function handleNewSymbol is used as a helper function in readTree function below.
+// Stack countOfChildren is used to control the count of children of all vertices.
+// At the bottom is the count of children of the root.
+// Stack treeTracing is used to trace the way back to the root of the tree.
+// vertexCounter is the number which is assigned as a key to a new node.
+// Symbol is the character extracted of the stream.
 void handleNewSymbol(std::stack<unsigned short>& countOfChildren, std::stack<Node*>& treeTracing, unsigned& vertexCounter, char symbol) {
     
     if (countOfChildren.top() == 0) {
@@ -84,7 +99,10 @@ void handleNewSymbol(std::stack<unsigned short>& countOfChildren, std::stack<Nod
     ++vertexCounter;
 }
 
-void readTree(Node*& root, std::ifstream& stream, unsigned& vertexCounter)
+// readTree extracts characters of the stream and constructs a tree with nodes in which data the characters are saved.
+// vertexCounter is passed by reference, because during the construction of the tree it is incremented every
+// time a new node is created.
+bool readTree(Node*& root, std::ifstream& stream, unsigned& vertexCounter)
 {
     std::stack<Node*> treeTracing;
     std::stack<unsigned short> countOfChildren;
@@ -94,7 +112,7 @@ void readTree(Node*& root, std::ifstream& stream, unsigned& vertexCounter)
     stream >> symbol;
     
     if (stream.eof())
-        return;
+        return 0;
     
     root = new Node(vertexCounter++, symbol);
     treeTracing.push(root);
@@ -104,7 +122,7 @@ void readTree(Node*& root, std::ifstream& stream, unsigned& vertexCounter)
     {
         stream >> symbol;
         if (stream.eof())
-            return;
+            break;
         
         if (symbol >= 'a' && symbol <= 'z') {
             handleNewSymbol(countOfChildren, treeTracing, vertexCounter, symbol);
@@ -122,19 +140,24 @@ void readTree(Node*& root, std::ifstream& stream, unsigned& vertexCounter)
             continue;
         }
     }
+    return 1;
 }
 
-void loadTreeFromFile(Node*& root, unsigned& vertexCounter)
+bool loadTreeFromFile(Node*& root, unsigned& vertexCounter)
 {
     std::ifstream stream("ternary_tree.txt");
     if (!stream) {
         std::cerr << "Cannot open file for reading\n";
-        return;
+        return 0;
     }
     
-    readTree(root, stream, vertexCounter);
-    
+    if (readTree(root, stream, vertexCounter)) {
+        stream.close();
+        return 1;
+    }
     stream.close();
+    return 0;
+
 }
 
 void deleteTree(Node* root) {
@@ -150,10 +173,10 @@ int main() {
     
     Node* root = nullptr;
     unsigned vertexCounter = 0;
-    loadTreeFromFile(root, vertexCounter);
-    BFS(root, vertexCounter);
-    deleteTree(root);
-    
-    
+    if (loadTreeFromFile(root, vertexCounter)) {
+        BFS(root, vertexCounter);
+        deleteTree(root);
+    }
+
     return 0;
 }
